SCEP Renewal Tests
Test various conditions associated with SCEP renewals. Tests include erroneous scenarios such as renewal attempts with invalid certificate, or with tampered data, or even if the renewal is legitamate but is occuring too frequently.

These test do assume that a "ManagementCA" exists, and that the change recorded at Issue#419 has also been implemented.